requirements.txt
Flask==1.1.1
keras==2.8.0
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
plotly==5.7.0
scikit_learn==1.0.2
streamlit==1.8.1
tensorflow==2.8.0

python -m streamlit run Streamlit_app.py


